Update derivedDataCreatorDplusToPiKPi.cxx based on Vit's comment

+++ PWGHF/TableProducer/derivedDataCreatorDplusToPiKPi.cxx
@@ -129,7 +129,7 @@ struct HfDerivedDataCreatorDplusToPiKPi {
   }
 
   template <typename T>
-  void fillTablesCandidate(const T& candidate, int candFlag, double invMass,
+  void fillTablesCandidate(const T& candidate, uint8_t candFlag, double invMass,
                            double ct, double y, int8_t flagMc, int8_t origin, int8_t swapping, int8_t flagDecayChan, const std::vector<float>& mlScores)
   {
     rowsCommon.fillTablesCandidate(candidate, invMass, y);
